Restructuring of project
