+++ src/App.js
@@ -1,7 +1,8 @@
 import React, { Component } from 'react';
 import classes from './App.module.css';
-import Person from '../src/components/Person/Person';
+import Persons from './components/Persons/Persons'
 import ErrorBoundary from './components/ErrorBoundary/ErrorBoundary';
+import CockPit from '../src/components/Cockpit/Cockpit';
 
 
 class App extends Component {
@@ -42,42 +43,26 @@ class App extends Component {
   };
 
   render() {
-    const assignedClasses = [];
-    if (this.state.persons.length <= 2) {
-      assignedClasses.push(classes.red);
-    }
-    if (this.state.persons.length <= 1) {
-      assignedClasses.push(classes.bold);
-    }
-
     let persons = null;
-    let btnClass = classes.btn + ' ' + classes.btnNotToggle;
     if (this.state.showPersons) {
       persons = (
         <div>
-          {
-            // Outputting List of persons
-            this.state.persons.map((person, index) => {
-              return <ErrorBoundary key={person.id}>
-                <Person
-                  name={person.name}
-                  age={person.age}
-                  click={()=> this.deletePersonHandler(index)}
-                  change={(event)=>this.nameChangeHandler(event, person.id)}
-                />
-              </ErrorBoundary>
-            })
-          }
+          <Persons 
+            persons = {this.state.persons}
+            clicked = {this.deletePersonHandler}
+            changed = {this.nameChangeHandler}
+          />
         </div>
       )
-      btnClass = classes.btn + ' ' + classes.btnToggle;;
     }
 
     return (
       <div className={classes.App}>
-        <h1>Hi, I'm a React App</h1>
-        <p className={assignedClasses.join(' ')}>This is really working!</p>
-        <div><button className={btnClass} onClick={this.togglePersonsHandler}>Toggle Names</button></div>
+        <CockPit
+          showPersons = {this.state.showPersons}
+          persons = {this.state.persons}
+          clicked = {this.togglePersonsHandler}
+        />
         {persons}
       </div>
     );
